Removed ConferenceSyncHelper.kt
