+++ apps/student/src/test/java/com/instructure/student/test/conferences/conference_list/ConferenceListNetworkDataSourceTest.kt
@@ -17,13 +17,15 @@
 
 package com.instructure.student.test.conferences.conference_list
 
+import com.instructure.canvasapi2.apis.ConferencesApi
 import com.instructure.canvasapi2.apis.OAuthAPI
 import com.instructure.canvasapi2.builders.RestParams
 import com.instructure.canvasapi2.models.AuthenticatedSession
 import com.instructure.canvasapi2.models.CanvasContext
 import com.instructure.canvasapi2.models.Conference
+import com.instructure.canvasapi2.models.ConferenceList
 import com.instructure.canvasapi2.utils.DataResult
-import com.instructure.pandautils.features.offline.sync.ConferenceSyncHelper
+import com.instructure.canvasapi2.utils.LinkHeaders
 import com.instructure.student.mobius.conferences.conference_list.datasource.ConferenceListNetworkDataSource
 import io.mockk.coEvery
 import io.mockk.coVerify
@@ -37,35 +39,76 @@ import org.junit.Test
 @ExperimentalCoroutinesApi
 class ConferenceListNetworkDataSourceTest {
 
-    private val conferenceSyncHelper: ConferenceSyncHelper = mockk(relaxed = true)
+    private val conferencesApi: ConferencesApi.ConferencesInterface = mockk(relaxed = true)
     private val oAuthApi: OAuthAPI.OAuthInterface = mockk(relaxed = true)
     private lateinit var networkDataSource: ConferenceListNetworkDataSource
 
     @Before
     fun setup() {
-        networkDataSource = ConferenceListNetworkDataSource(conferenceSyncHelper, oAuthApi)
+        networkDataSource = ConferenceListNetworkDataSource(conferencesApi, oAuthApi)
     }
 
     @Test
     fun `Return conferences list api model`() = runTest {
-        val expected = DataResult.Success(listOf(Conference(1), Conference(2)))
-        coEvery { conferenceSyncHelper.getConferencesForContext(any(), any()) } returns expected
+        val expected = listOf(Conference(1), Conference(2))
+        coEvery { conferencesApi.getConferencesForContext(any(), any()) } returns DataResult.Success(ConferenceList(expected))
 
         val result = networkDataSource.getConferencesForContext(CanvasContext.emptyCourseContext(1), true)
 
-        TestCase.assertEquals(expected, result)
-        coVerify(exactly = 1) { conferenceSyncHelper.getConferencesForContext(CanvasContext.emptyCourseContext(1), true) }
+        TestCase.assertEquals(DataResult.Success(expected), result)
+        coVerify(exactly = 1) {
+            conferencesApi.getConferencesForContext(
+                CanvasContext.emptyCourseContext(1).toAPIString().drop(1),
+                RestParams(isForceReadFromNetwork = true)
+            )
+        }
+    }
+
+    @Test
+    fun `Depaginate conferences`() = runTest {
+        val page1 = ConferenceList(listOf(Conference(1), Conference(2), Conference(3)))
+        val page2 = ConferenceList(listOf(Conference(4), Conference(5), Conference(6)))
+        val page3 = ConferenceList(listOf(Conference(7), Conference(8), Conference(9)))
+
+        coEvery { conferencesApi.getConferencesForContext(any(), any()) } returns DataResult.Success(
+            page1,
+            linkHeaders = LinkHeaders(nextUrl = "page_2_url")
+        )
+
+        coEvery { conferencesApi.getNextPage("page_2_url", any()) } returns DataResult.Success(
+            page2,
+            linkHeaders = LinkHeaders(nextUrl = "page_3_url")
+        )
+
+        coEvery { conferencesApi.getNextPage("page_3_url", any()) } returns DataResult.Success(page3)
+
+        val result = networkDataSource.getConferencesForContext(CanvasContext.emptyCourseContext(1), true)
+
+        TestCase.assertEquals(DataResult.Success(page1.conferences + page2.conferences + page3.conferences), result)
+        coVerify(exactly = 1) {
+            conferencesApi.getConferencesForContext(
+                CanvasContext.emptyCourseContext(1).toAPIString().drop(1),
+                RestParams(isForceReadFromNetwork = true)
+            )
+            conferencesApi.getNextPage("page_2_url", RestParams(isForceReadFromNetwork = true))
+            conferencesApi.getNextPage("page_3_url", RestParams(isForceReadFromNetwork = true))
+        }
     }
 
     @Test
     fun `Return failed result if conferences call fails`() = runTest {
         val expected = DataResult.Fail()
-        coEvery { conferenceSyncHelper.getConferencesForContext(any(), any()) } returns expected
+        coEvery { conferencesApi.getConferencesForContext(any(), any()) } returns expected
 
         val result = networkDataSource.getConferencesForContext(CanvasContext.emptyCourseContext(1), true)
 
         TestCase.assertEquals(expected, result)
-        coVerify(exactly = 1) { conferenceSyncHelper.getConferencesForContext(CanvasContext.emptyCourseContext(1), true) }
+        coVerify(exactly = 1) {
+            conferencesApi.getConferencesForContext(
+                CanvasContext.emptyCourseContext(1).toAPIString().drop(1),
+                RestParams(isForceReadFromNetwork = true)
+            )
+        }
     }
 
     @Test
