+++ libs/pandautils/src/main/java/com/instructure/pandautils/features/offline/sync/OfflineSyncWorker.kt
@@ -25,7 +25,9 @@ import com.instructure.canvasapi2.apis.*
 import com.instructure.canvasapi2.builders.RestParams
 import com.instructure.canvasapi2.models.AssignmentGroup
 import com.instructure.canvasapi2.models.CanvasContext
+import com.instructure.canvasapi2.models.Conference
 import com.instructure.canvasapi2.models.ScheduleItem
+import com.instructure.canvasapi2.utils.DataResult
 import com.instructure.canvasapi2.utils.depaginate
 import com.instructure.pandautils.room.offline.daos.*
 import com.instructure.pandautils.room.offline.entities.CourseSettingsEntity
@@ -33,6 +35,7 @@ import com.instructure.pandautils.room.offline.entities.DashboardCardEntity
 import com.instructure.pandautils.room.offline.entities.PageEntity
 import com.instructure.pandautils.room.offline.entities.QuizEntity
 import com.instructure.pandautils.room.offline.facade.AssignmentFacade
+import com.instructure.pandautils.room.offline.facade.ConferenceFacade
 import com.instructure.pandautils.room.offline.facade.CourseFacade
 import com.instructure.pandautils.room.offline.facade.ScheduleItemFacade
 import dagger.assisted.Assisted
@@ -57,7 +60,8 @@ class OfflineSyncWorker @AssistedInject constructor(
     private val dashboardCardDao: DashboardCardDao,
     private val courseSettingsDao: CourseSettingsDao,
     private val scheduleItemFacade: ScheduleItemFacade,
-    private val conferenceSyncHelper: ConferenceSyncHelper
+    private val conferencesApi: ConferencesApi.ConferencesInterface,
+    private val conferenceFacade: ConferenceFacade
 ) : CoroutineWorker(context, workerParameters) {
 
     override suspend fun doWork(): Result {
@@ -86,7 +90,7 @@ class OfflineSyncWorker @AssistedInject constructor(
                         syllabusCourseIds.add(courseSettings.courseId)
                     }
                     if (courseSettings.conferences) {
-                        conferenceSyncHelper.fetchConferences(courseSettings.courseId)
+                        fetchConferences(courseSettings.courseId)
                     }
                 }
             fetchSyllabus(syllabusCourseIds)
@@ -190,6 +194,23 @@ class OfflineSyncWorker @AssistedInject constructor(
                     quiz?.let { quizDao.insert(QuizEntity(it)) }
                 }
             }
+        }
+    }
+
+    private suspend fun fetchConferences(courseId: Long) {
+        val conferences = getConferencesForContext(CanvasContext.emptyCourseContext(courseId), true).dataOrNull
+        conferences?.let { conferenceFacade.insertConferences(it, courseId) }
+    }
+
+    private suspend fun getConferencesForContext(
+        canvasContext: CanvasContext, forceNetwork: Boolean
+    ): DataResult<List<Conference>> {
+        val params = RestParams(isForceReadFromNetwork = forceNetwork)
+
+        return conferencesApi.getConferencesForContext(canvasContext.toAPIString().drop(1), params).map {
+            it.conferences
+        }.depaginate { url ->
+            conferencesApi.getNextPage(url, params).map { it.conferences }
         }
     }
 }
